+++ src/test/java/com/codeborne/selenide/impl/TailOfCollectionTest.java
@@ -13,10 +13,11 @@ class TailOfCollectionTest implements WithAssertions {
   private final WebElement element1 = mock(WebElement.class);
   private final WebElement element2 = mock(WebElement.class);
   private final WebElement element3 = mock(WebElement.class);
-  private WebElementsCollection originalCollection = mock(WebElementsCollection.class);
+  private final WebElementsCollection originalCollection = mock(WebElementsCollection.class);
 
   @BeforeEach
   void setUp() {
+    when(originalCollection.description()).thenReturn("li.active");
     when(originalCollection.getElements()).thenReturn(asList(element1, element2, element3));
   }
 
@@ -39,5 +40,12 @@ class TailOfCollectionTest implements WithAssertions {
     TailOfCollection $$ = new TailOfCollection(originalCollection, 4);
     assertThat($$.getElements())
       .isEqualTo(asList(element1, element2, element3));
+  }
+
+  @Test
+  void description() {
+    TailOfCollection $$ = new TailOfCollection(originalCollection, 4);
+    assertThat($$.description())
+      .isEqualTo("li.active:last(4)");
   }
 }